+++ src/constants/StateKeys.ts
@@ -5,4 +5,5 @@ export const StateKeys = {
   },
   executingDemoFile: "executingDemoFile",
   variables: "variables",
+  presentationView: "presentationView",
 };