+++ src/preview/components/SlideControl.tsx
@@ -1,0 +1,33 @@
+import * as React from 'react';
+import { Icon } from 'vscrui';
+
+export interface ISlideControlProps {
+  action: () => void;
+  iconName?: string;
+  icon?: React.ReactNode;
+  title: string;
+  disabled?: boolean;
+  isSlideControl?: boolean;
+}
+
+export const SlideControl: React.FunctionComponent<ISlideControlProps> = ({
+  action,
+  iconName,
+  icon,
+  title,
+  disabled,
+  isSlideControl,
+}: React.PropsWithChildren<ISlideControlProps>) => {
+  return (
+    <button
+      title={title}
+      onClick={action}
+      disabled={disabled}
+      className={`p-2 inline-flex justify-center items-center rounded hover:bg-[var(--vscode-toolbar-hoverBackground)] disabled:opacity-50 disabled:cursor-not-allowed ${!isSlideControl ? ' opacity-70 hover:opacity-100' : ''}`}
+    >
+      <span className="sr-only">{title}</span>
+      {iconName && <Icon name={iconName as any} className="!text-[var(--vscode-editorWidget-foreground)] inline-flex justify-center items-center" />}
+      {icon}
+    </button>
+  );
+};