+++ src/utils/index.ts
@@ -38,6 +38,7 @@ export * from "./replaceContent";
 export * from "./sanitizeFileName";
 export * from "./setContext";
 export * from "./sleep";
+export * from "./togglePresentationView";
 export * from "./updateConfig";
 export * from "./upperCaseFirstLetter";
 export * from "./writeFile";